10/12/2020

+++ AZTU-Akademik/Models/Announcement.cs
@@ -5,7 +5,7 @@ namespace AZTU_Akademik.Models
 {
     public partial class Announcement
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public DateTime? CreateDate { get; set; }
